refactored layout and header
